+++ src/Generators/Tests/DelegateTestGenerator.php
@@ -15,6 +15,7 @@ use PhpUnitGen\Core\Contracts\Generators\TestGenerator;
 use PhpUnitGen\Core\Exceptions\RuntimeException;
 use PhpUnitGen\Core\Generators\Tests\Basic\BasicTestGenerator;
 use PhpUnitGen\Core\Generators\Tests\Laravel\Channel\ChannelTestGenerator;
+use PhpUnitGen\Core\Generators\Tests\Laravel\Command\CommandTestGenerator;
 use PhpUnitGen\Core\Generators\Tests\Laravel\LaravelTestGenerator;
 use PhpUnitGen\Core\Generators\Tests\Laravel\Policy\PolicyTestGenerator;
 use PhpUnitGen\Core\Helpers\Str;
@@ -92,6 +93,10 @@ class DelegateTestGenerator implements DelegateTestGeneratorContract, ConfigAwar
         if ($this->isLaravelProject()) {
             if (Str::contains('\\Broadcasting\\', $reflectionClass->getName())) {
                 return ChannelTestGenerator::class;
+            }
+
+            if (Str::contains('\\Console\\Commands\\', $reflectionClass->getName())) {
+                return CommandTestGenerator::class;
             }
 
             if (Str::contains('\\Policies\\', $reflectionClass->getName())) {