+++ tests/Unit/Generators/Tests/DelegateTestGeneratorTest.php
@@ -29,6 +29,8 @@ use PhpUnitGen\Core\Generators\Tests\Basic\BasicMethodFactory;
 use PhpUnitGen\Core\Generators\Tests\Basic\BasicTestGenerator;
 use PhpUnitGen\Core\Generators\Tests\DelegateTestGenerator;
 use PhpUnitGen\Core\Generators\Tests\Laravel\Channel\ChannelTestGenerator;
+use PhpUnitGen\Core\Generators\Tests\Laravel\Command\CommandTestGenerator;
+use PhpUnitGen\Core\Generators\Tests\Laravel\FeatureClassFactory;
 use PhpUnitGen\Core\Generators\Tests\Laravel\LaravelTestGenerator;
 use PhpUnitGen\Core\Generators\Tests\Laravel\Policy\PolicyMethodFactory;
 use PhpUnitGen\Core\Generators\Tests\Laravel\Policy\PolicyTestGenerator;
@@ -230,6 +232,60 @@ class DelegateTestGeneratorTest extends TestCase
                     && $config->implementations() === [
                         TestGenerator::class                => ChannelTestGenerator::class,
                         ClassFactoryContract::class         => UnitClassFactory::class,
+                        DocumentationFactoryContract::class => DocumentationFactory::class,
+                        ImportFactoryContract::class        => ImportFactory::class,
+                        MethodFactory::class                => BasicMethodFactory::class,
+                        PropertyFactoryContract::class      => PropertyFactory::class,
+                        StatementFactoryContract::class     => StatementFactory::class,
+                        ValueFactoryContract::class         => ValueFactory::class,
+                    ];
+            }))
+            ->willReturn($container);
+
+        $container->shouldReceive('get')
+            ->with(TestGenerator::class)
+            ->andReturn($delegatedTestGenerator);
+
+        $delegatedTestGenerator->shouldReceive('generate')
+            ->with($reflectionClass)
+            ->andReturn($testClass);
+
+        $this->assertSame($testClass, $this->testGenerator->generate($reflectionClass));
+    }
+
+    public function testGenerateDelegatesToLaravelCommand(): void
+    {
+        $reflectionClass = Mockery::mock(ReflectionClass::class);
+        $container = Mockery::mock(ContainerInterface::class);
+        $delegatedTestGenerator = Mockery::mock(TestGenerator::class);
+        $testClass = Mockery::mock(TestClass::class);
+
+        $reflectionClass->shouldReceive([
+            'isInterface' => false,
+            'isAnonymous' => false,
+            'getName'     => 'App\\Console\\Commands\\ProductCommand',
+        ]);
+
+        $this->config->shouldReceive([
+            'toArray' => [
+                'automaticGeneration' => false,
+                'implementations'     => [
+                    TestGenerator::class => DelegateTestGenerator::class,
+                    MethodFactory::class => BasicMethodFactory::class,
+                ],
+            ],
+        ]);
+        $this->config->shouldReceive('getOption')
+            ->with('context')
+            ->andReturn('laravel');
+
+        $this->testGenerator->expects($this->once())
+            ->method('makeNewContainer')
+            ->with($this->callback(function (ConfigContract $config) {
+                return $config->automaticGeneration() === false
+                    && $config->implementations() === [
+                        TestGenerator::class                => CommandTestGenerator::class,
+                        ClassFactoryContract::class         => FeatureClassFactory::class,
                         DocumentationFactoryContract::class => DocumentationFactory::class,
                         ImportFactoryContract::class        => ImportFactory::class,
                         MethodFactory::class                => BasicMethodFactory::class,